refactor out blindfold mode with Cline
